pyproject.toml
[tool.poetry]
name = "tft-ui"
version = "0.1.0"
description = "Testing Farm UI."
authors = ["Jan Havlin <jhavlin@redhat.com>"]
readme = "README.adoc"

packages = [
  { include = "tft", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
reflex = "^0.6"
uvicorn =  {extras = ["standard"], version = "^0.20.0"}
pyjwt = "^2.8.0"
dynaconf = "^3.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true
